+++ EssentialFeedJorgeiOSTests/Controllers/FeedViewControllerTests.swift
@@ -363,68 +363,6 @@ final class FeedViewControllerTests: XCTestCase {
 }
 
 
-// MARK: - Extension. FeedViewController
-private extension FeedViewController {
-    var isShowingLoadingIndicator: Bool {
-        return refreshControl?.isRefreshing == true
-    }
-    
-    
-    func simulateUserInitiatedFeedReload() {
-        refreshControl?.simulatePullToRefresh()
-    }
-    
-    
-    func simulateFeedImageViewNearVisible(at row: Int) {
-        let ds = tableView.prefetchDataSource
-        let index = IndexPath(row: row, section: feedImagesSection)
-        ds?.tableView(tableView, prefetchRowsAt: [index])
-    }
-    
-    
-    func simulateFeedImageViewNotNearVisible(at row: Int) {
-        simulateFeedImageViewNearVisible(at: row)
-        
-        let ds = tableView.prefetchDataSource
-        let index = IndexPath(row: row, section: feedImagesSection)
-        ds?.tableView?(tableView, cancelPrefetchingForRowsAt: [index])
-    }
-    
-    
-    private var feedImagesSection: Int {
-        return 0
-    }
-    
-    
-    @discardableResult
-    func simulateFeedImageViewVisible(at index: Int) -> FeedImageTableViewCell? {
-        return feedImageView(at: index) as? FeedImageTableViewCell
-    }
-    
-    
-    func simulateFeedImageViewNotVisible(at row: Int) {
-        let view = simulateFeedImageViewVisible(at: row)
-        
-        let delegate = tableView.delegate
-        let index = IndexPath(row: row, section: feedImagesSection)
-        delegate?.tableView?(tableView, didEndDisplaying: view!, forRowAt: index)
-    }
-    
-    
-    func numberOfRenderedFeedImageViews() -> Int {
-        return tableView.numberOfRows(inSection: feedImagesSection)
-    }
-    
-    
-    func feedImageView(at row: Int) -> UITableViewCell? {
-        let ds = tableView.dataSource
-        let index = IndexPath(row: row, section: feedImagesSection)
-        
-        return ds?.tableView(tableView, cellForRowAt: index)
-    }
-}
-
-
 // MARK: - Extension. FeedImageTableViewCell
 private extension FeedImageTableViewCell {
     var isShowingLocation: Bool {
