+++ EssentialFeedJorgeiOSTests/Helpers/FeedViewController+TestHelpers.swift
@@ -1,0 +1,72 @@
+//
+//  FeedViewController+TestHelpers.swift
+//  EssentialFeedJorgeiOSTests
+//
+//  Created by jfuentescasillas on 15/09/23.
+//
+
+
+import UIKit
+import EssentialFeedJorgeiOS
+
+
+// MARK: - Extension. FeedViewController
+extension FeedViewController {
+    var isShowingLoadingIndicator: Bool {
+        return refreshControl?.isRefreshing == true
+    }
+    
+    
+    func simulateUserInitiatedFeedReload() {
+        refreshControl?.simulatePullToRefresh()
+    }
+    
+    
+    func simulateFeedImageViewNearVisible(at row: Int) {
+        let ds = tableView.prefetchDataSource
+        let index = IndexPath(row: row, section: feedImagesSection)
+        ds?.tableView(tableView, prefetchRowsAt: [index])
+    }
+    
+    
+    func simulateFeedImageViewNotNearVisible(at row: Int) {
+        simulateFeedImageViewNearVisible(at: row)
+        
+        let ds = tableView.prefetchDataSource
+        let index = IndexPath(row: row, section: feedImagesSection)
+        ds?.tableView?(tableView, cancelPrefetchingForRowsAt: [index])
+    }
+    
+    
+    private var feedImagesSection: Int {
+        return 0
+    }
+    
+    
+    @discardableResult
+    func simulateFeedImageViewVisible(at index: Int) -> FeedImageTableViewCell? {
+        return feedImageView(at: index) as? FeedImageTableViewCell
+    }
+    
+    
+    func simulateFeedImageViewNotVisible(at row: Int) {
+        let view = simulateFeedImageViewVisible(at: row)
+        
+        let delegate = tableView.delegate
+        let index = IndexPath(row: row, section: feedImagesSection)
+        delegate?.tableView?(tableView, didEndDisplaying: view!, forRowAt: index)
+    }
+    
+    
+    func numberOfRenderedFeedImageViews() -> Int {
+        return tableView.numberOfRows(inSection: feedImagesSection)
+    }
+    
+    
+    func feedImageView(at row: Int) -> UITableViewCell? {
+        let ds = tableView.dataSource
+        let index = IndexPath(row: row, section: feedImagesSection)
+        
+        return ds?.tableView(tableView, cellForRowAt: index)
+    }
+}
